+++ cmd/root/kubernetes/kubernetes.go
@@ -94,7 +94,7 @@ func (c *KubectlK8sActions) CopyFromHost(hostString string, source, destination 
 	if err != nil {
 		return "", fmt.Errorf("unable to get container name: %v", err)
 	}
-	return c.cli.Execute(false, c.kubectlPath, "cp", "-n", c.namespace, "-c", container, fmt.Sprintf("%v:%v", hostString, source), c.cleanLocal(destination))
+	return c.cli.Execute(false, c.kubectlPath, "cp", "-n", c.namespace, "-c", container, "--retries", "5", fmt.Sprintf("%v:%v", hostString, source), c.cleanLocal(destination))
 }
 
 func (c *KubectlK8sActions) CopyToHost(hostString string, source, destination string) (out string, err error) {
@@ -107,7 +107,7 @@ func (c *KubectlK8sActions) CopyToHost(hostString string, source, destination st
 	if err != nil {
 		return "", fmt.Errorf("unable to get container name: %v", err)
 	}
-	return c.cli.Execute(false, c.kubectlPath, "cp", "-n", c.namespace, "-c", container, c.cleanLocal(source), fmt.Sprintf("%v:%v", hostString, destination))
+	return c.cli.Execute(false, c.kubectlPath, "cp", "-n", c.namespace, "-c", container, "--retries", "5", c.cleanLocal(source), fmt.Sprintf("%v:%v", hostString, destination))
 }
 
 func (c *KubectlK8sActions) GetCoordinators() (podName []string, err error) {